add chain signature algorithm recognition + refactoring
