+++ plugin/payroll/constants.go
@@ -1,0 +1,12 @@
+package payroll
+
+const PLUGIN_TYPE = "payroll"
+const erc20ABI = `[{
+    "name": "transfer",
+    "type": "function",
+    "inputs": [
+        {"name": "recipient", "type": "address"},
+        {"name": "amount", "type": "uint256"}
+    ],
+    "outputs": [{"name": "", "type": "bool"}]
+}]`
