+++ plugin/payroll/policy.go
@@ -1,0 +1,139 @@
+package payroll
+
+import (
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"math/big"
+	"strings"
+
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	gcommon "github.com/ethereum/go-ethereum/common"
+	gtypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/rlp"
+	"github.com/sirupsen/logrus"
+	"github.com/vultisig/vultisigner/internal/types"
+)
+
+func (p *PayrollPlugin) ValidateTransactionProposal(policy types.PluginPolicy, txs []types.PluginKeysignRequest) error {
+	err := p.ValidatePluginPolicy(policy)
+	if err != nil {
+		return fmt.Errorf("failed to validate plugin policy: %v", err)
+	}
+
+	parsedABI, err := abi.JSON(strings.NewReader(erc20ABI))
+	if err != nil {
+		return fmt.Errorf("failed to parse ABI: %v", err)
+	}
+
+	var payrollPolicy types.PayrollPolicy
+	if err := json.Unmarshal(policy.Policy, &payrollPolicy); err != nil {
+		return fmt.Errorf("fail to unmarshal payroll policy, err: %w", err)
+	}
+
+	for i, tx := range txs {
+		var parsedTx *gtypes.Transaction
+		txBytes, err := hex.DecodeString(tx.Transaction)
+		if err != nil {
+			return fmt.Errorf("failed to decode transaction: %v", err)
+		}
+
+		err = rlp.DecodeBytes(txBytes, &parsedTx)
+		if err != nil {
+			return fmt.Errorf("failed to parse transaction: %v", err)
+		}
+
+		txDestination := parsedTx.To()
+		if txDestination == nil {
+			return fmt.Errorf("transaction destination is nil")
+		}
+
+		if strings.ToLower(txDestination.Hex()) != strings.ToLower(payrollPolicy.TokenID[i]) { //todo : why we compare to tokenId and not recipient address?
+			return fmt.Errorf("transaction destination does not match token ID")
+		}
+
+		txData := parsedTx.Data()
+		m, err := parsedABI.MethodById(txData[:4])
+		if err != nil {
+			return fmt.Errorf("failed to get method by ID: %v", err)
+		}
+
+		v := make(map[string]interface{})
+		if err := m.Inputs.UnpackIntoMap(v, txData[4:]); err != nil {
+			return fmt.Errorf("failed to unpack transaction data: %v", err)
+		}
+
+		fmt.Printf("Decoded: %+v\n", v)
+
+		recipientAddress, ok := v["recipient"].(gcommon.Address)
+		if !ok {
+			return fmt.Errorf("failed to get recipient address")
+		}
+
+		var recipientFound bool
+		for _, recipient := range payrollPolicy.Recipients {
+			if strings.EqualFold(recipientAddress.Hex(), recipient.Address) {
+				recipientFound = true
+				break
+			}
+		}
+
+		if !recipientFound {
+			return fmt.Errorf("recipient not found in policy")
+		}
+	}
+
+	return nil
+}
+
+func (p *PayrollPlugin) ValidatePluginPolicy(policyDoc types.PluginPolicy) error {
+	if policyDoc.PluginType != PLUGIN_TYPE {
+		return fmt.Errorf("policy does not match plugin type, expected: %s, got: %s", PLUGIN_TYPE, policyDoc.PluginType)
+	}
+
+	var payrollPolicy types.PayrollPolicy
+	if err := json.Unmarshal(policyDoc.Policy, &payrollPolicy); err != nil {
+		return fmt.Errorf("fail to unmarshal payroll policy, err: %w", err)
+	}
+
+	if len(payrollPolicy.Recipients) == 0 {
+		return fmt.Errorf("no recipients found in payroll policy")
+	}
+
+	if len(payrollPolicy.ChainID) != len(payrollPolicy.Recipients) {
+		p.logger.WithFields(logrus.Fields{
+			"chain_id_length":   payrollPolicy.ChainID,
+			"recipients_length": payrollPolicy.Recipients,
+		}).Error("chain_id array length does not match number of recipients")
+		return fmt.Errorf("chain_id array length must match number of recipients")
+	}
+
+	if len(payrollPolicy.TokenID) != len(payrollPolicy.Recipients) {
+		return fmt.Errorf("token_id array length must match number of recipients")
+	}
+
+	for _, recipient := range payrollPolicy.Recipients {
+		mixedCaseAddress, err := gcommon.NewMixedcaseAddressFromString(recipient.Address)
+		if err != nil {
+			return fmt.Errorf("invalid recipient address: %s", recipient.Address)
+		}
+
+		// if the address is not all lowercase, check the checksum
+		if strings.ToLower(recipient.Address) != recipient.Address {
+			if !mixedCaseAddress.ValidChecksum() {
+				return fmt.Errorf("invalid recipient address checksum: %s", recipient.Address)
+			}
+		}
+
+		if recipient.Amount == "" {
+			return fmt.Errorf("amount is required for recipient %s", recipient.Address)
+		}
+
+		_, ok := new(big.Int).SetString(recipient.Amount, 10)
+		if !ok {
+			return fmt.Errorf("invalid amount for recipient %s: %s", recipient.Address, recipient.Amount)
+		}
+	}
+
+	return nil
+}
